+++ src/storage.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use crate::value::Something;
 
 pub struct Object {
-    properties: HashMap<u32, Something>,
+    properties: HashMap<u64, Something>,
     references: u32,
 }
 
@@ -34,23 +34,23 @@ impl Object {
         return self.references > 0;
     }
 
-    pub fn set_property(&mut self, key: u32, value: Something) -> Option<Something> {
+    pub fn set_property(&mut self, key: u64, value: Something) -> Option<Something> {
         return self.properties.insert(key, value);
     }
 
-    pub fn get_property(&self, key: u32) -> Option<&Something> {
+    pub fn get_property(&self, key: u64) -> Option<&Something> {
         return self.properties.get(&key);
     }
 
-    pub fn delete_property(&mut self, key: u32) -> Option<Something> {
+    pub fn delete_property(&mut self, key: u64) -> Option<Something> {
         self.properties.remove(&key)
     }
 }
 
 pub struct Storage {
-    pub collection: HashMap<u32, Object>,
-    pub blobs: HashMap<u32, Vec<u8>>,
-    last_id: u32,
+    pub collection: HashMap<u64, Object>,
+    pub blobs: HashMap<u64, Vec<u8>>,
+    last_id: u64,
 }
 
 impl Storage {
@@ -62,7 +62,7 @@ impl Storage {
         }
     }
 
-    pub fn try_drop(&mut self, id: u32) -> Option<()> {
+    pub fn try_drop(&mut self, id: u64) -> Option<()> {
         let obj = self.collection.get_mut(&id)?;
         obj.decrement_references();
         if !obj.has_references() {
@@ -77,30 +77,30 @@ impl Storage {
         return Some(());
     }
 
-    pub fn get_reference_count(&self, id: u32) -> Option<u32> {
+    pub fn get_reference_count(&self, id: u64) -> Option<u32> {
         let obj = self.collection.get(&id)?;
         return Some(obj.references);
     }
 
-    pub fn get_blob_pointer(&self, id: u32) -> Option<(*const u8, usize)> {
+    pub fn get_blob_pointer(&self, id: u64) -> Option<(*const u8, usize)> {
         let blob = self.blobs.get(&id)?;
         return Some((blob.as_ptr(), blob.len()));
     }
 
-    pub fn add_blob(&mut self, data: Vec<u8>) -> u32 {
+    pub fn add_blob(&mut self, data: Vec<u8>) -> u64 {
         let id = self.last_id;
         self.last_id += 1;
         self.blobs.insert(id, data);
         return id;
     }
 
-    pub fn increment_object_references(&mut self, id: u32) -> Option<bool> {
+    pub fn increment_object_references(&mut self, id: u64) -> Option<bool> {
         let obj = self.collection.get_mut(&id)?;
         obj.increment_references();
         return Some(true);
     }
 
-    pub fn create_object(&mut self, kind: ObjectKind) -> u32 {
+    pub fn create_object(&mut self, kind: ObjectKind) -> u64 {
         let id = match kind {
             ObjectKind::Object => self.object_id(),
             ObjectKind::Array => self.array_id(),
@@ -109,11 +109,11 @@ impl Storage {
         return id;
     }
 
-    pub fn get_object(&self, id: u32) -> Option<&Object> {
+    pub fn get_object(&self, id: u64) -> Option<&Object> {
         return self.collection.get(&id);
     }
 
-    pub fn set_object_property(&mut self, object_id: u32, key: u32, value: Something) {
+    pub fn set_object_property(&mut self, object_id: u64, key: u64, value: Something) {
         if let Something::Ref(to) = value {
             if let Some(obj) = self.collection.get_mut(&to) {
                 obj.increment_references();
@@ -126,14 +126,14 @@ impl Storage {
         }
     }
 
-    pub fn get_object_property(&self, object_id: u32, key: u32) -> Option<&Something> {
+    pub fn get_object_property(&self, object_id: u64, key: u64) -> Option<&Something> {
         if let Some(object) = self.collection.get(&object_id) {
             return object.get_property(key);
         }
         return None;
     }
 
-    pub fn delete_object_property(&mut self, object_id: u32, key: u32) -> Option<Something> {
+    pub fn delete_object_property(&mut self, object_id: u64, key: u64) -> Option<Something> {
         let object = self.collection.get_mut(&object_id)?;
         let prop = object.delete_property(key)?;
         if let Something::Ref(to) = prop {
@@ -143,13 +143,13 @@ impl Storage {
         return Some(prop);
     }
 
-    fn object_id(&mut self) -> u32 {
+    fn object_id(&mut self) -> u64 {
         let id = self.last_id;
         self.last_id += 1;
         return id << 1;
     }
 
-    fn array_id(&mut self) -> u32 {
+    fn array_id(&mut self) -> u64 {
         let id = self.last_id;
         self.last_id += 1;
         return (id << 1) | 0b1;
